+++ Form/PersonType.php
@@ -1,0 +1,56 @@
+<?php
+
+/*
+ * This file is part of aakb/kontrolgruppen-core-bundle.
+ *
+ * (c) 2019 ITK Development
+ *
+ * This source file is subject to the MIT license.
+ */
+
+namespace Kontrolgruppen\CoreBundle\Form;
+
+use Kontrolgruppen\CoreBundle\Entity\Person;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+/**
+ * Class PersonType.
+ */
+class PersonType extends AbstractType
+{
+    /**
+     * @param FormBuilderInterface $builder
+     * @param array                $options
+     */
+    public function buildForm(FormBuilderInterface $builder, array $options)
+    {
+        $builder
+            ->add('CPR', TextType::class, [
+                'label' => 'client.form.person.cpr',
+                'empty_data' => '',
+            ])
+            ->add('name', TextType::class, [
+                'label' => 'client.form.person.name',
+                'empty_data' => '',
+            ])
+            ->add('highlighted', CheckboxType::class, [
+                'label' => 'client.form.person.highlighted',
+                'required' => false,
+            ])
+        ;
+    }
+
+    /**
+     * @param OptionsResolver $resolver
+     */
+    public function configureOptions(OptionsResolver $resolver)
+    {
+        $resolver->setDefaults([
+            'data_class' => Person::class,
+        ]);
+    }
+}