+++ Repository/PersonRepository.php
@@ -1,0 +1,32 @@
+<?php
+
+/*
+ * This file is part of aakb/kontrolgruppen-core-bundle.
+ *
+ * (c) 2019 ITK Development
+ *
+ * This source file is subject to the MIT license.
+ */
+
+namespace Kontrolgruppen\CoreBundle\Repository;
+
+use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
+use Doctrine\Common\Persistence\ManagerRegistry;
+use Kontrolgruppen\CoreBundle\Entity\Person;
+
+/**
+ * @method Person|null find($id, $lockMode = null, $lockVersion = null)
+ * @method Person|null findOneBy(array $criteria, array $orderBy = null)
+ * @method Person[]    findAll()
+ * @method Person[]    findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
+ */
+class PersonRepository extends ServiceEntityRepository
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function __construct(ManagerRegistry $registry)
+    {
+        parent::__construct($registry, Person::class);
+    }
+}