+++ Service/ProcessManager.php
@@ -12,13 +12,11 @@ namespace Kontrolgruppen\CoreBundle\Service;
 
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
-use Kontrolgruppen\CoreBundle\CPR\Cpr;
-use Kontrolgruppen\CoreBundle\CPR\CprException;
 use Kontrolgruppen\CoreBundle\CPR\CprServiceInterface;
-use Kontrolgruppen\CoreBundle\Entity\Client;
+use Kontrolgruppen\CoreBundle\Entity\AbstractProcessClient;
 use Kontrolgruppen\CoreBundle\Entity\Conclusion;
 use Kontrolgruppen\CoreBundle\Entity\Process;
-use Kontrolgruppen\CoreBundle\Entity\ProcessType;
+use Kontrolgruppen\CoreBundle\Entity\ProcessClientPerson;
 use Kontrolgruppen\CoreBundle\Entity\User;
 use Kontrolgruppen\CoreBundle\Repository\ProcessRepository;
 use Psr\Log\LoggerInterface;
@@ -141,20 +139,15 @@ class ProcessManager
     /**
      * Create new process.
      *
-     * @param Process|null     $process
-     * @param ProcessType|null $processType
+     * @param Process               $process
+     * @param AbstractProcessClient $client
      *
      * @return \Kontrolgruppen\CoreBundle\Entity\Process
      *
      * @throws \Exception
      */
-    public function newProcess(Process $process = null, ProcessType $processType = null)
+    public function newProcess(Process $process, AbstractProcessClient $client = null)
     {
-        if (null === $process) {
-            $process = new Process();
-            $process->setProcessType($processType);
-        }
-
         $resourceToLock = 'case-number';
 
         $this->lockService->createLock($resourceToLock);
@@ -168,7 +161,7 @@ class ProcessManager
         $process->setCaseNumber($this->getNewCaseNumber());
         $process->setProcessStatus($this->decideStatusForProcess($process));
         $process->setConclusion($this->createConclusionForProcess($process));
-        $process->setClient($this->createClientForProcess($process));
+        $process->setProcessClient($client);
 
         $this->entityManager->persist($process);
         $this->entityManager->flush();
@@ -222,18 +215,35 @@ class ProcessManager
         $process->setCompletedAt($completedAt);
         $process->setLastCompletedAt($completedAt);
 
-        $calculatedRevenue = $this->economyService->calculateRevenue($process);
-        $netCollectiveSum = $calculatedRevenue['netCollectiveSum'] ?: null;
+        if ($this->isRevenueAvailable($process)) {
+            $calculatedRevenue = $this->economyService->calculateRevenue($process);
+            $netCollectiveSum = $calculatedRevenue['netCollectiveSum'] ?: null;
 
-        if (!empty($process->getLastNetCollectiveSum())) {
-            $netCollectiveSumDifference = $netCollectiveSum - $process->getLastNetCollectiveSum();
-            $process->setNetCollectiveSumDifference($netCollectiveSumDifference);
+            if (!empty($process->getLastNetCollectiveSum())) {
+                $netCollectiveSumDifference = $netCollectiveSum - $process->getLastNetCollectiveSum();
+                $process->setNetCollectiveSumDifference($netCollectiveSumDifference);
+            }
+
+            $process->setLastNetCollectiveSum($netCollectiveSum);
         }
-
-        $process->setLastNetCollectiveSum($netCollectiveSum);
 
         $this->entityManager->persist($process);
         $this->entityManager->flush();
+    }
+
+    /**
+     * Decide if revenue is available for a process, i.e. if it makes sense to
+     * calculate revenue for the process.
+     *
+     * @param Process|null $process
+     *
+     * @return bool
+     */
+    public function isRevenueAvailable(Process $process = null): bool
+    {
+        return null !== $process
+            && null !== $process->getProcessClient()
+            && ProcessClientPerson::PERSON === $process->getProcessClient()->getType();
     }
 
     /**
@@ -273,23 +283,5 @@ class ProcessManager
         $conclusionClass = $process->getProcessType()->getConclusionClass();
 
         return new $conclusionClass();
-    }
-
-    /**
-     * @param Process $process
-     *
-     * @return Client
-     */
-    private function createClientForProcess(Process $process): Client
-    {
-        $client = new Client();
-
-        try {
-            $client = $this->cprService->populateClient(new Cpr($process->getClientCPR()), $client);
-        } catch (CprException $e) {
-            $this->logger->error($e);
-        }
-
-        return $client;
     }
 }